fix: Add a client-side rate limiter
In other to prevent buggy scripts from hitting the server with too many
requests, add a rate limiter on the client side. With the settings in
the script, it's able to do ~ 20 requests in a second, and then it
continues making ~ 5 requests per second. If the script backs off, the
bucket gets refilled and the script is able to make another burst of
requests, so well behaved scripts shouldn't be impacted.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>